Moving the different plots to another folder
 - Trying to create folders for each integration.
 - New folder plotting for all the plots. As most of the plots are made from matplotlib, we will create different files for the different plots.
